\docType{package}
\name{rk.Teaching}
\alias{rk.Teaching}
\alias{rk.Teaching-package}
\title{rk.Teaching}
\description{
An RKWard plugin for teaching and learning Statistics.

\url{http://aprendeconalf.es/rkteaching/}
}
\details{
This package add a new menu \code{Teaching} to the RKWard GUI.
The new menu add procedures to perform basic statistics analysis.
The dialogs are designed to simplify the steps and options for the user focusing in the most important aspects of any procedure.

It contains procedures for: 
\itemize{ 
  \item Data manipulation: 
  \itemize{ 
    \item Data fitering 
    \item Calculation of variables 
    \item Recoding of variables 
    \item Data weighting 
    \item Standarization of variables
  }
  \item Frequency distribution:
  \itemize{
    \item Frequency tables 
    \item Two-dimensional frequency tables
  } 
  \item Charts: 
  \itemize{ 
    \item Bars chart 
    \item Histogram 
    \item Piechart 
    \item Box plot 
    \item Means chart 
    \item Interaction chart 
    \item Scatter plot 
    \item Scatter plot matrix 
  } 
  \item Descriptive statistics: 
  \itemize{
    \item Statistics 
  } 
  \item Regression: 
  \itemize{ 
    \item Linear regression 
    \item Non-linear regression 
    \item Model comparison 
    \item Predictions 
    \item Correlation 
  } 
  \item Parametric tests: 
  \itemize{ 
    \item Means: 
    \itemize{ 
      \item T-test for the mean of one population 
      \item T-test for the means of two independent populations
      \item T-test for the means of two paired populations
      \item ANOVA test 
      \item Sample size for estimating one mean
      \item Sample size for the T-test
    } 
    \item Variances: 
    \itemize{ 
      \item Fisher's test 
      \item Levene's test 
    } 
    \item Proportions: 
    \itemize{ 
      \item Test for one proportion
      \item Test for two proportions 
      \item Sample size for estimating one proportion
    } 
  } 
  \item Non-parametric tests: 
  \itemize{ 
    \item Normality: 
    \itemize{ 
      \item Lillieford (Komogorov-Smirnov) test 
      \item Shapiro-Wilk test 
    } 
    \item Mann-Whitney U test for two independent populations 
    \item Wilcoxon test for two paired populations
    \item Kruskal-Wallis test for several independent populations 
    \item Friedman test for several paired populations 
    \item Chi-square test of independence 
    \item McNemar test of independence
    \item Chi-square test of goodness of fit 
  } 
  \item Concordance:
  \itemize{ 
    \item Intraclass correlation coefficient 
    \item Cohen's kappa 
  } 
  \item Probability:
  \itemize{
    \item Gambling
    \item Build probability space
    \item Repeat probability space
    \item Join independent probability spaces
    \item Compute probability
  }
  \item Distributions: 
  \itemize{ 
    \item Discrete: 
    \itemize{ 
      \item Binomial
      \item Geometric
      \item Hypergeometric
      \item Poisson 
    } 
    \item Continuous: 
    \itemize{ 
      \item Uniform 
      \item Normal 
      \item Chi-square 
      \item Student's t 
      \item Fisher's F 
    } 
  } 
  \item Simulations: 
  \itemize{ 
    \item Law of rare events
    \item Central limit theorem
  }
}
}